Pack VCLS in a tar ball and pack MORE vcls
Still leaves the original mycat* work for legacy-support?

By using find, we also support deeper directory trees, which has been an
issue in the past.

Oh, and also fix a LONG standing issue with item numbering. For some reason
I originally incremented the item number between log* and echo* lines,
meaning that the first command would be stored as item 00, but listed as
item 01. This resolves that as a side effect.