selinux: Better capture AVC denials
The comm name is highly unreliable, it is supposed to be named after
the executable but can be at the thread granularity once we start naming
them, which we do in varnishd.

The varnishd semodule is also not limited to labeling varnishd, it also
has a context for varnishlog for example. So capturing any context that
matches varnish should bring us denials for both types varnishlog_t and
varnishd_t, and for the latter, denials originating from both mgt and
cache processes.