Add coredump analysis & debug mode
List coredumps of the Manager & child process of Varnish.
This contains useful information such as: time of the crash, pid, uid,
gid, signal, whether corefile is present, path of the executeable and size of
the corefile.

Gather further information of each coredump, such as but not limited to:
command line, cgroup, stack traces of threads, loaded so, path of the corefile.

Coredump analysis of the latest crash is done using
coredumpctl, we invoke gdb on the coredump to get the backtrace.
This depend on coredumpctl having --debugger-arguments.

A debug mode is introduced and activated through "-d", when activated we
invoke gdb on all Varnish instances to get full backtrace and threads
information. This is behind a non-default flag as it will pause Varnish while
gathering this information.

Signed-off-by: Asad Sajjad Ahmed <asadsa@varnish-software.com>